requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiogram==3.6.0
sqlalchemy[asyncio]==2.0.30
aiomysql==0.2.0
pydantic-settings==2.2.1
httpx==0.27.0
apscheduler==3.10.4
python-multipart==0.0.9
groq>=0.18.0
PyJWT>=2.8.0
alembic>=1.13.0